Software/RaspberryPi/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
attrs==23.2.0
    # via
    #   outcome
    #   trio
beautifulsoup4==4.10.0
    # via -r requirements.in
board==1.0
    # via -r requirements.in
certifi==2023.11.17
    # via
    #   requests
    #   selenium
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via
    #   pdfminer-six
    #   requests
cryptography==42.0.0
    # via pdfminer-six
exceptiongroup==1.2.0
    # via
    #   trio
    #   trio-websocket
furl==2.1.3
    # via -r requirements.in
h11==0.14.0
    # via wsproto
idna==3.6
    # via
    #   requests
    #   trio
neopixel==0.0.1
    # via -r requirements.in
numpy==1.26.3
    # via
    #   neopixel
    #   pandas
    #   scipy
orderedmultidict==1.0.1
    # via furl
outcome==1.3.0.post0
    # via trio
pandas==2.2.0
    # via -r requirements.in
pdfminer-six==20221105
    # via pdfplumber
pdfplumber==0.10.3
    # via -r requirements.in
pillow==10.2.0
    # via pdfplumber
pyaudio==0.2.14
    # via neopixel
pycparser==2.21
    # via cffi
pypdfium2==4.26.0
    # via pdfplumber
pyserial==3.5
    # via -r requirements.in
pysocks==1.7.1
    # via urllib3
python-dateutil==2.8.2
    # via pandas
pytz==2023.3.post1
    # via pandas
requests==2.31.0
    # via -r requirements.in
scipy==1.12.0
    # via neopixel
selenium==4.17.2
    # via -r requirements.in
six==1.16.0
    # via
    #   furl
    #   orderedmultidict
    #   python-dateutil
sniffio==1.3.0
    # via trio
sortedcontainers==2.4.0
    # via trio
soupsieve==2.5
    # via beautifulsoup4
trio==0.24.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.11.1
    # via selenium
typing-extensions==4.9.0
    # via selenium
tzdata==2023.4
    # via pandas
urllib3[socks]==2.1.0
    # via
    #   requests
    #   selenium
wsproto==1.2.0
    # via trio-websocket
